requirements.txt
networkx
pandas
colomoto_jupyter
tqdm
bonesis==0.6.5
